requirements.txt
selenium
webdriver-manager
telethon
pyperclip
python-dotenv
